move a bunch of cpp/h code files into cppForSwig/Utils/
