+++ cppForSwig/Utils/ArmoryConfig.cpp
