+++ cppForSwig/Utils/ArmoryConfig.h
