+++ cppForSwig/Utils/BIP150_151.cpp
