+++ cppForSwig/Utils/BIP15x_Handshake.h
