+++ cppForSwig/Utils/BinaryData.cpp
