+++ cppForSwig/Utils/BinaryData.h
