+++ cppForSwig/Utils/BitcoinSettings.cpp
