+++ cppForSwig/Utils/BitcoinSettings.h
