+++ cppForSwig/Utils/BtcUtils.cpp
