+++ cppForSwig/Utils/BtcUtils.h
