+++ cppForSwig/Utils/Cryptography.cpp
