+++ cppForSwig/Utils/Cryptography.h
