+++ cppForSwig/Utils/DBUtils.cpp
