+++ cppForSwig/Utils/HttpMessage.cpp
