+++ cppForSwig/Utils/HttpMessage.h
