+++ cppForSwig/Utils/JSON_codec.cpp
