+++ cppForSwig/Utils/SecureBinaryData.cpp
