+++ cppForSwig/Utils/SecureBinaryData.h
