+++ cppForSwig/Utils/ThreadSafeClasses.h
