+++ cppForSwig/Utils/log.cpp
