+++ cppForSwig/Utils/log.h
