+++ cppForSwig/deprecated/PartialMerkle.h
