+++ cppForSwig/deprecated/old_not_very_good_tests.cpp
