+++ cppForSwig/deprecated/playground.cpp
