+++ cppForSwig/deprecated/util.h
